requirements.txt
asgiref==3.2.3
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
Django==3.0.1
django-cors-headers==3.2.0
mysqlclient==1.4.6
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
gunicorn==20.0.0

